Moved examples into librf24 folder
